New parameters for beatmatching, extra options for bandpass filter, crush, and coarse
I’ve added the new parameters “matchcps” and “stretchTo” that simplify
beatmatching or automatically stretching samples to specified lengths
with Tidal.  These override the “speed “ and “accelerate” parameters.

I’ve also added some extra functionality to the bandpass filter and the
quantizers “coarse” and “crush”.  Using a negative frequency in “bandf”
will switch the bandpass to a bandnotch filter.  Using a negative value
for “crush” will use a nonlinear quantization scheme that drops the
effective noise floor a bit and handles quiet samples at low bit depths
a bit better.  And using a negative value for “coarse” will use
different “resampling” method that has a bit less aliasing and
effectively low-passes the sound.

Finally, to try and future-proof Dirt a little, I also made a slight
change to the way the server handles messages with too many parameters.
 Instead of logging a generic error and playing no sound, it will now
log a shorter error message but play the sound if the rest of the
parameters are valid.

This updated Dirt should be fully backward compatible with prior
versions of Tidal.